requirements.txt
# =============================================================================
# Real Estate Bot - Dependencias
# =============================================================================
# Instalar con: pip install -r requirements.txt

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Selenium (para sitios con JavaScript)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Base de datos
# SQLite viene incluido en Python

# Configuración
PyYAML>=6.0
python-dotenv>=1.0.0

# Notificaciones
python-telegram-bot>=20.0

# Utilidades
urllib3>=2.0.0

# Opcional: Scrapy para proyectos más avanzados
# scrapy>=2.11.0

# Desarrollo
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
